# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: string;
  name: string;
  description: string;
  price: number;
  isVeg: boolean;
  isAvailable: boolean;
  emoji: string;
}

export interface MenuCategory {
  id: string;
  name: string;
  items: MenuItem[];
}

export interface OrderItem {
  id: string;
  name: string;
  price: number;
  quantity: number;
  instructions: string;
  isVeg: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  preparing: 'preparing',
  ready: 'ready',
  served: 'served',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  tableNumber: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  createdAt: string;
  estimatedMinutes: number;
}

export interface OrderInput {
  tableNumber: number;
  items: OrderItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  placed: 'placed',
  preparing: 'preparing',
  ready: 'ready',
  served: 'served',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  estimatedMinutes?: number;
}

export type WaiterCallType = typeof WaiterCallType[keyof typeof WaiterCallType];


export const WaiterCallType = {
  waiter: 'waiter',
  water: 'water',
  bill: 'bill',
} as const;

export type WaiterCallStatus = typeof WaiterCallStatus[keyof typeof WaiterCallStatus];


export const WaiterCallStatus = {
  pending: 'pending',
  claimed: 'claimed',
  resolved: 'resolved',
} as const;

export interface WaiterCall {
  id: string;
  tableNumber: number;
  type: WaiterCallType;
  status: WaiterCallStatus;
  createdAt: string;
}

export type CallInputType = typeof CallInputType[keyof typeof CallInputType];


export const CallInputType = {
  waiter: 'waiter',
  water: 'water',
  bill: 'bill',
} as const;

export interface CallInput {
  tableNumber: number;
  type: CallInputType;
}

export type CallStatusUpdateStatus = typeof CallStatusUpdateStatus[keyof typeof CallStatusUpdateStatus];


export const CallStatusUpdateStatus = {
  pending: 'pending',
  claimed: 'claimed',
  resolved: 'resolved',
} as const;

export interface CallStatusUpdate {
  status: CallStatusUpdateStatus;
}

export interface DashboardSummary {
  activeOrders: number;
  pendingCalls: number;
  freeWaiters: number;
  busyWaiters: number;
}