# lib/api-zod/src/generated/types/callInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallInputType } from './callInputType';

export interface CallInput {
  tableNumber: number;
  type: CallInputType;
}