# lib/api-zod/src/generated/types/callStatusUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallStatusUpdateStatus } from './callStatusUpdateStatus';

export interface CallStatusUpdate {
  status: CallStatusUpdateStatus;
}