# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: string;
  name: string;
  price: number;
  quantity: number;
  instructions: string;
  isVeg: boolean;
}