# lib/api-zod/src/generated/types/orderStatusUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusUpdateStatus } from './orderStatusUpdateStatus';

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  estimatedMinutes?: number;
}