# lib/api-zod/src/generated/types/waiterCall.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WaiterCallStatus } from './waiterCallStatus';
import type { WaiterCallType } from './waiterCallType';

export interface WaiterCall {
  id: string;
  tableNumber: number;
  type: WaiterCallType;
  status: WaiterCallStatus;
  createdAt: string;
}